Address MR comments

+++ db/post_migrate/20210629101712_remove_deprecated_modsecurity_columns.rb
@@ -8,15 +8,13 @@ class RemoveDeprecatedModsecurityColumns < ActiveRecord::Migration[6.0]
   INDEX_NAME = 'index_clusters_applications_ingress_on_modsecurity'
 
   def up
-    with_lock_retries do
-      remove_column :clusters_applications_ingress, :modsecurity_enabled
-      remove_column :clusters_applications_ingress, :modsecurity_mode
-    end
+    remove_column :clusters_applications_ingress, :modsecurity_enabled if column_exists?(:clusters_applications_ingress, :modsecurity_enabled)
+    remove_column :clusters_applications_ingress, :modsecurity_mode if column_exists?(:clusters_applications_ingress, :modsecurity_mode)
   end
 
   def down
-    add_column :clusters_applications_ingress, :modsecurity_enabled, :boolean
-    add_column :clusters_applications_ingress, :modsecurity_mode, :smallint, null: false, default: 0
+    add_column :clusters_applications_ingress, :modsecurity_enabled, :boolean unless column_exists?(:clusters_applications_ingress, :modsecurity_enabled)
+    add_column :clusters_applications_ingress, :modsecurity_mode, :smallint, null: false, default: 0 unless column_exists?(:clusters_applications_ingress, :modsecurity_mode)
 
     add_concurrent_index :clusters_applications_ingress, [:modsecurity_enabled, :modsecurity_mode, :cluster_id], name: INDEX_NAME
   end
